Refactor for best practice.
Changed up return code to be more managable with remote powershell execution and better pipeline behavior. This script will now return a "Completed" field which is a Boolean, which returns based on whether you have any sophos components left installed, and (if there are any) a LeftoverPrograms field, which is a hash table, which contains the Get-WMIObject query output of the leftover programs.